Move make_combined_wave to tabor specific code and simplify it; Cache hash of TaborSegement
