+++ tests/_program/tabor_tests.py
@@ -3,10 +3,15 @@ import itertools
 import numpy as np
 from unittest import mock
 
+try:
+    import pytabor
+except ImportError:
+    pytabor = None
+
 from teawg import model_properties_dict
 
 from qupulse._program.tabor import TaborException, TaborProgram, \
-    TaborSegment, TaborSequencing, PlottableProgram, TableDescription
+    TaborSegment, TaborSequencing, PlottableProgram, TableDescription, make_combined_wave
 from qupulse._program._loop import MultiChannelProgram, Loop
 from qupulse._program.instructions import InstructionBlock
 from qupulse.hardware.util import voltage_to_uint16
@@ -14,6 +19,8 @@ from qupulse.utils.types import TimeType
 
 from tests.pulses.sequencing_dummies import DummyWaveform
 from tests._program.loop_tests import LoopTests, WaveformGenerator, MultiChannelTests
+
+from tests.hardware import dummy_modules
 
 
 class PlottableProgramTests(unittest.TestCase):
@@ -493,3 +500,116 @@ class TaborSegmentTests(unittest.TestCase):
         for seg_a, seg_b in itertools.product(all_segments, all_segments):
             if seg_a == seg_b:
                 self.assertEqual(hash(seg_a), hash(seg_b))
+
+
+class TaborMakeCombinedTest(unittest.TestCase):
+    @staticmethod
+    def validate_result(tabor_segments, result, fill_value=None):
+        pos = 0
+        for i, tabor_segment in enumerate(tabor_segments):
+            if i > 0:
+                if tabor_segment.ch_b is None:
+                    if fill_value:
+                        np.testing.assert_equal(result[pos:pos + 16],
+                                                np.full(16, fill_value=fill_value, dtype=np.uint16))
+                else:
+                    np.testing.assert_equal(result[pos:pos + 16], np.full(16, tabor_segment.ch_b[0], dtype=np.uint16))
+                pos += 16
+
+                if tabor_segment.ch_a is None:
+                    if fill_value:
+                        np.testing.assert_equal(result[pos:pos + 16],
+                                                np.full(16, fill_value=fill_value, dtype=np.uint16))
+                else:
+                    np.testing.assert_equal(result[pos:pos + 16], np.full(16, tabor_segment.ch_a[0], dtype=np.uint16))
+                pos += 16
+
+            for j in range(tabor_segment.num_points // 16):
+                if tabor_segment.ch_b is None:
+                    if fill_value:
+                        np.testing.assert_equal(result[pos:pos + 16],
+                                                np.full(16, fill_value=fill_value, dtype=np.uint16))
+                else:
+                    np.testing.assert_equal(result[pos:pos + 16], tabor_segment.ch_b[j * 16: (j + 1) * 16])
+                pos += 16
+
+                if tabor_segment.ch_a is None:
+                    if fill_value:
+                        np.testing.assert_equal(result[pos:pos + 16],
+                                                np.full(16, fill_value=fill_value, dtype=np.uint16))
+                else:
+                    np.testing.assert_equal(result[pos:pos + 16], tabor_segment.ch_a[j * 16: (j + 1) * 16])
+                pos += 16
+
+    def exec_general(self, data_1, data_2):
+        tabor_segments = [TaborSegment.from_sampled(d1, d2, None, None) for d1, d2 in zip(data_1, data_2)]
+        expected_length = (sum(segment.num_points for segment in tabor_segments) + 16 * (len(tabor_segments) - 1)) * 2
+
+        result = make_combined_wave(tabor_segments)
+        self.assertEqual(len(result), expected_length)
+
+        self.validate_result(tabor_segments, result)
+
+        destination_array = np.empty(expected_length, dtype=np.uint16)
+        result = make_combined_wave(tabor_segments, destination_array=destination_array)
+        self.validate_result(tabor_segments, result)
+        self.assertEqual(destination_array.data, result.data)
+
+    def test_make_comb_both(self):
+        gen = itertools.count()
+        data_1 = [np.fromiter(gen, count=32, dtype=np.uint16),
+                  np.fromiter(gen, count=16, dtype=np.uint16),
+                  np.fromiter(gen, count=192, dtype=np.uint16)]
+
+        data_2 = [np.fromiter(gen, count=32, dtype=np.uint16),
+                  np.fromiter(gen, count=16, dtype=np.uint16),
+                  np.fromiter(gen, count=192, dtype=np.uint16)]
+        for d in data_2:
+            d += 1000
+
+        self.exec_general(data_1, data_2)
+
+    def test_make_single_chan(self):
+        gen = itertools.count()
+        data_1 = [np.fromiter(gen, count=32, dtype=np.uint16),
+                  np.fromiter(gen, count=16, dtype=np.uint16),
+                  np.fromiter(gen, count=192, dtype=np.uint16)]
+
+        data_2 = [None]*len(data_1)
+        self.exec_general(data_1, data_2)
+        self.exec_general(data_2, data_1)
+
+    def test_empty_segment_list(self):
+        combined = make_combined_wave([])
+
+        self.assertIsInstance(combined, np.ndarray)
+        self.assertIs(combined.dtype, np.dtype('uint16'))
+        self.assertEqual(len(combined), 0)
+
+
+@unittest.skipIf(pytabor in (dummy_modules.dummy_pytabor, None), "Cannot compare to pytabor results")
+class TaborMakeCombinedPyTaborCompareTest(TaborMakeCombinedTest):
+    def exec_general(self, data_1, data_2, fill_value=None):
+        tabor_segments = [TaborSegment.from_sampled(d1, d2, None, None) for d1, d2 in zip(data_1, data_2)]
+        expected_length = (sum(segment.num_points for segment in tabor_segments) + 16 * (len(tabor_segments) - 1)) * 2
+
+        offset = 0
+        pyte_result = 15000*np.ones(expected_length, dtype=np.uint16)
+        for i, segment in enumerate(tabor_segments):
+            offset = pytabor.make_combined_wave(segment.ch_a, segment.ch_b,
+                                                dest_array=pyte_result, dest_array_offset=offset,
+                                                add_idle_pts=i > 0)
+        self.assertEqual(expected_length, offset)
+
+        result = make_combined_wave(tabor_segments)
+        np.testing.assert_equal(pyte_result, result)
+
+        dest_array = 15000*np.ones(expected_length, dtype=np.uint16)
+        result = make_combined_wave(tabor_segments, destination_array=dest_array)
+        np.testing.assert_equal(pyte_result, result)
+        # test that the destination array data is not copied
+        self.assertEqual(dest_array.__array_interface__['data'],
+                         result.__array_interface__['data'])
+
+        with self.assertRaises(ValueError):
+            make_combined_wave(tabor_segments, destination_array=np.ones(16))
